@@ -122,24 +122,25 @@
   },
   "homepage": "https://github.com/graphistry/viz-app#readme",
   "dependencies": {
-    "@graphistry/client-api": "~2.12.3",
-    "@graphistry/common": "3.1.2",
+    "@graphistry/client-api": "~2.12.6",
+    "@graphistry/common": "~3.2.4",
     "@graphistry/config": "3.0.6",
     "@graphistry/dijkstra": "2.0.0",
-    "@graphistry/falcor": "~2.9.4",
-    "@graphistry/falcor-json-graph": "2.8.0",
-    "@graphistry/falcor-model-rxjs": "~2.9.4",
-    "@graphistry/falcor-path-syntax": "2.8.0",
-    "@graphistry/falcor-path-utils": "~2.9.4",
-    "@graphistry/falcor-query-syntax": "~2.9.4",
-    "@graphistry/falcor-react-redux": "~2.9.4",
-    "@graphistry/falcor-react-schema": "~2.9.4",
-    "@graphistry/falcor-router": "~2.9.4",
-    "@graphistry/falcor-router-saddle": "~2.9.4",
-    "@graphistry/falcor-socket-datasource": "~2.9.4",
+    "@graphistry/falcor": "~2.9.6",
+    "@graphistry/falcor-json-graph": "~2.9.6",
+    "@graphistry/falcor-model-rxjs": "~2.9.6",
+    "@graphistry/falcor-path-syntax": "~2.9.6",
+    "@graphistry/falcor-path-utils": "~2.9.6",
+    "@graphistry/falcor-query-syntax": "~2.9.6",
+    "@graphistry/falcor-react-redux": "~2.9.6",
+    "@graphistry/falcor-react-schema": "~2.9.6",
+    "@graphistry/falcor-router": "~2.9.6",
+    "@graphistry/falcor-router-saddle": "~2.9.6",
+    "@graphistry/falcor-socket-datasource": "~2.9.6",
     "@graphistry/node-pigz": "2.0.0",
     "@graphistry/rc-slider": "5.1.3",
     "@graphistry/react-split-pane": "^0.1.63",
+    "@graphistry/vgraph-to-mapd": "~1.1.1",
     "babel-core": "6.24.1",
     "babel-runtime": "6.23.0",
     "bcrypt": "1.0.2",
@@ -189,31 +190,30 @@
     "pegjs": "0.10.0",
     "pegjs-util": "1.4.3",
     "prop-types": "~15.5.10",
-    "protobufjs": "5.0.0",
     "q": "1.5.0",
     "raf": "3.3.2",
     "rc-color-picker": "1.1.4",
     "rc-switch": "1.4.4",
-    "react": "15.5.4",
+    "react": "~15.6.1",
     "react-ace": "4.2.3",
     "react-bootstrap": "0.31.0",
     "react-dock": "0.2.4",
-    "react-dom": "15.5.4",
-    "react-helmet": "5.0.3",
+    "react-dom": "~15.6.1",
+    "react-helmet": "~5.1.3",
     "react-measure": "1.4.7",
     "react-overlays": "0.7.0",
-    "react-redux": "5.0.4",
+    "react-redux": "^5.0.5",
     "react-resize-detector": "0.5.0",
     "react-router": "4.1.1",
     "react-router-dom": "4.1.1",
     "react-select": "1.0.0-rc.2",
-    "react-virtualized": "9.7.5",
+    "react-virtualized": "~9.8.0",
     "recompose": "0.23.4",
-    "redbox-react": "1.3.6",
-    "redux": "3.6.0",
+    "redbox-react": "~1.4.2",
+    "redux": "~3.7.1",
     "redux-observable": "0.14.1",
     "resize-observer-polyfill": "1.4.2",
-    "rxjs": "5.4.0",
+    "rxjs": "~5.4.2",
     "rxjs-gestures": "1.1.8",
     "sanitize-html": "1.14.1",
     "simpleflakes": "1.2.3",
@@ -289,7 +289,6 @@
     "postcss-font-awesome": "0.3.4",
     "postcss-loader": "2.0.5",
     "progress-bar-webpack-plugin": "1.9.3",
-    "proto-loader": "0.2.0",
     "react-addons-perf": "15.4.2",
     "react-addons-test-utils": "15.5.1",
     "react-hot-loader": "3.0.0-beta.6",
