@@ -8,8 +8,7 @@ const sprintf = require('sprintf-js').sprintf;
 const log         = require('@graphistry/common').logger;
 const logger      = log.createLogger('graph-viz:data:vgraphwriter');
 
-const ProtoBuf = require('protobufjs/dist/protobuf-light');
-const protoBufDefinitions = ProtoBuf.loadJson(require('viz-app/vgraph/graph_vector.proto')).build();
+import { VectorGraph } from '@graphistry/vgraph-to-mapd/lib/cjs/vgraph';
 
 /* Hack way to serialize positions while waiting for dataframe */
 function serializePositions(graph) {
@@ -29,16 +28,16 @@ function serializePositions(graph) {
             yVal[i] = values[2*i + 1];
         }
 
-        const xVec = new protoBufDefinitions.VectorGraph.DoubleAttributeVector();
+        const xVec = new VectorGraph.DoubleAttributeVector();
         xVec.name = 'x';
         xVec.values = xVal;
-        xVec.target = protoBufDefinitions.VectorGraph.AttributeTarget.VERTEX;
+        xVec.target = VectorGraph.AttributeTarget.VERTEX;
         vg.double_vectors.push(xVec);
 
-        const yVec = new protoBufDefinitions.VectorGraph.DoubleAttributeVector();
+        const yVec = new VectorGraph.DoubleAttributeVector();
         yVec.name = 'y';
         yVec.values = yVal;
-        yVec.target = protoBufDefinitions.VectorGraph.AttributeTarget.VERTEX;
+        yVec.target = VectorGraph.AttributeTarget.VERTEX;
         vg.double_vectors.push(yVec);
 
         return vg;
@@ -49,7 +48,7 @@ function save(graph, name) {
     logger.debug('Saving current graph as', name);
 
     return serializePositions(graph).then((vg) => {
-        const blob = vg.encode().toBuffer();
+        const blob = VectorGraph.encode(vg).finish();
         logger.debug('Uploading to S3', name);
         return s3.upload(config.S3, config.BUCKET, {name: name}, blob, {ContentEncoding: 'gzip'});
     }).fail(log.makeQErrorHandler(logger, 'save vgraph'));
